fix(pty): share submitted input filtering
